fix test error
