+++ extra-requirements.txt
@@ -1,1 +1,2 @@
 msgpack>=0.5.6
+PyMySQL>=1.0.2
