--- requirements.txt
@@ -1,1 +1,0 @@
-PyMySQL>=1.0.2
