+++ tox.ini
@@ -4,6 +4,8 @@ minversion = 2.0
 skipsdist = True
 envlist = py27, py34, py35, py36, py37, pep8, cover
 deps = -r{toxinidir}/test-requirements.txt
+       -r{toxinidir}/extra-requirements.txt
+       -r{toxinidir}/requirements.txt
 
 [testenv]
 
@@ -11,6 +13,8 @@ setenv = VIRTUAL_ENV={envdir}
 
 usedevelop = True
 deps = -r{toxinidir}/test-requirements.txt
+       -r{toxinidir}/extra-requirements.txt
+       -r{toxinidir}/requirements.txt
 whitelist_externals =
   bash
   find
